requirements.txt

numpy>=1.20
pytest>=6.0
PyYAML>=5.4
networkx>=2.6
matplotlib>=3.4
